scripts/requirements-silver.txt
# AI Employee - Silver Tier Dependencies
# Install with: pip install -r requirements-silver.txt

# Core dependencies (Bronze Tier)
watchdog>=4.0.0          # File system monitoring for watchers

# Gmail API (Silver Tier)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# WhatsApp Web Automation (Silver Tier)
playwright>=1.40.0

# LinkedIn API (Silver Tier)
linkedin-api>=2.0.0
requests>=2.31.0

# Email MCP Server
python-dotenv>=1.0.0
# smtplib is built-in to Python

# Scheduling (Silver Tier)
schedule>=1.2.0
apscheduler>=3.10.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
